Moved simple test (not involving ES) to dedicated dirs for travis build.
