+++ search/src/test/java/org/ensembl/genesearch/output/QueryOutputTest.java
@@ -12,7 +12,7 @@
  *  limitations under the License.
  */
 
-package org.ensembl.genesearch;
+package org.ensembl.genesearch.output;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -24,6 +24,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.ensembl.genesearch.QueryOutput;
 import org.junit.Test;
 
 /**
